agent-gemini3/pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = [
    { name = "Zheng Liu", email = "zheng.liu@outlook.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.0",
    "python-dotenv>=1.0.1",
    "langchain",
    "langchain-openai",
    "pydantic",
    "pytest",
    "pytest-asyncio",
    "langchain-core>=1.1.0",
    "aiohttp>=3.13.2",
    "langchain-mcp-adapters>=0.1.14",
    "rich>=14.2.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.7",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
    "debugpy>=1.8.14",
]
